requirements.txt
flask
flask-cors
numpy
pandas
langchain
langchain_experimental
python-dotenv
langchain-google-genai
kmodes
kneed
mlxtend
torch